Bugs fixed

+++ SmartEventPlatformWeb/Controllers/LocationsController.cs
@@ -37,7 +37,7 @@ namespace SmartEventPlatformWeb.Controllers
                 return NotFound();
             }
 
-            var locationDetails = await _context.Locations
+            var vm = await _context.Locations
                 .Where(l => l.LocationId == id)
                 .Select(l => new LocationDetailsViewModel
                 {
@@ -46,13 +46,13 @@ namespace SmartEventPlatformWeb.Controllers
                     Address = l.Address,
                     Capacity = l.Capacity
                 }).FirstOrDefaultAsync();
-            
-            if (locationDetails == null)
+
+            if (vm == null)
             {
                 return NotFound();
             }
 
-            return View(locationDetails);
+            return View(vm);
         }
 
         public IActionResult Create()
@@ -64,19 +64,19 @@ namespace SmartEventPlatformWeb.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(LocationCreateViewModel vm)
         {
-            if (ModelState.IsValid)
-            {   
-                var location = new Location
-                {
-                    LocationName = vm.LocationName,
-                    Address = vm.Address,
-                    Capacity = vm.Capacity
-                };
-                _context.Add(location);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+            if (!ModelState.IsValid)
+            {
+                return View(vm);
             }
-            return View(vm);
+            var location = new Location
+            {
+                LocationName = vm.LocationName,
+                Address = vm.Address,
+                Capacity = vm.Capacity
+            };
+            _context.Add(location);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
         }
 
         public async Task<IActionResult> Edit(long? id)
@@ -110,37 +110,37 @@ namespace SmartEventPlatformWeb.Controllers
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
-                try
-                {
-                    var location = await _context.Locations.FindAsync(id);
-                    if(location == null)
-                    {
-                        return NotFound();
-                    }
-
-                    location.LocationName = vm.LocationName;
-                    location.Address = vm.Address;
-                    location.Capacity = vm.Capacity;
-
-                    _context.Update(location);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!LocationExists(vm.LocationId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                return View(vm);
             }
-            return View(vm);
+            try
+            {
+                var location = await _context.Locations.FindAsync(id);
+                if (location == null)
+                {
+                    return NotFound();
+                }
+
+                location.LocationName = vm.LocationName;
+                location.Address = vm.Address;
+                location.Capacity = vm.Capacity;
+
+                _context.Update(location);
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!LocationExists(vm.LocationId))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+            return RedirectToAction(nameof(Index));
         }
 
         public async Task<IActionResult> Delete(long? id)
@@ -179,6 +179,13 @@ namespace SmartEventPlatformWeb.Controllers
                 return NotFound();
             }
 
+            if (await LocationHasDependenciesAsync(id))
+            {
+                return ReturnLocationDeleteViewWithError(
+                    location,
+                    "This location cannot be deleted because it is used by one or more events.");
+            }
+
             try
             {
                 _context.Locations.Remove(location);
@@ -188,24 +195,41 @@ namespace SmartEventPlatformWeb.Controllers
             }
             catch (DbUpdateException)
             {
-                ModelState.AddModelError(string.Empty,
+                return ReturnLocationDeleteViewWithError(
+                    location,
                     "This location cannot be deleted because it is used by one or more events.");
-
-                var vm = new LocationDeleteViewModel
-                {
-                    LocationId = location.LocationId,
-                    LocationName = location.LocationName,
-                    Address = location.Address,
-                    Capacity = location.Capacity
-                };
-
-                return View("Delete", vm);
             }
         }
 
         private bool LocationExists(long id)
         {
             return _context.Locations.Any(l => l.LocationId == id);
+        }
+
+        private async Task<bool> LocationHasDependenciesAsync(long locationId)
+        {
+            return await _context.Events
+                .AnyAsync(e => e.LocationId == locationId);
+        }
+
+        private static LocationDeleteViewModel MapToDeleteViewModel(Location location)
+        {
+            return new LocationDeleteViewModel
+            {
+                LocationId = location.LocationId,
+                LocationName = location.LocationName,
+                Address = location.Address,
+                Capacity = location.Capacity
+            };
+        }
+
+        private IActionResult ReturnLocationDeleteViewWithError(Location location, string errorMessage)
+        {
+            ModelState.AddModelError(string.Empty, errorMessage);
+
+            var vm = MapToDeleteViewModel(location);
+
+            return View("Delete", vm);
         }
     }
 }
